requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
prophet
tensorflow
keras
dotenv
PyYAML